Refactor availability classes to use long for occupied and capacity values

+++ travel-agency-query-side/src/main/java/com/rzodeczko/domain/model/AvailabilityStatusPolicy.java
@@ -7,7 +7,7 @@ public record AvailabilityStatusPolicy(double lastRoomsThreshold) {
         }
     }
 
-    public AvailabilityStatus evaluate(int occupied, int capacity) {
+    public AvailabilityStatus evaluate(long occupied, long capacity) {
         if (occupied >= capacity) {
             return AvailabilityStatus.SOLD_OUT;
         }
